+++ src/strojure/undertow/handler/websocket.clj
@@ -1,20 +1,13 @@
 (ns strojure.undertow.handler.websocket
   "WebSocket handler functionality."
-  (:require [strojure.undertow.api.types :as types])
-  (:import (clojure.lang IPersistentMap)
-           (io.undertow.websockets WebSocketConnectionCallback)
+  (:import (io.undertow.websockets WebSocketConnectionCallback)
            (io.undertow.websockets.core WebSocketChannel)
            (io.undertow.websockets.spi WebSocketHttpExchange)
-           (org.xnio ChannelListener)
-           (strojure.undertow.websocket WebSocketChannelListener)))
+           (org.xnio ChannelListener)))
 
 (set! *warn-on-reflection* true)
 
 ;;,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-
-(defmethod types/as-websocket-listener IPersistentMap
-  [config]
-  (WebSocketChannelListener. config))
 
 (defn listener-as-connection-callback
   "Returns connection callback for given `ChannelListener`."